C3: Catch errors in quarantined e2e tests
