+++ packages/create-cloudflare/e2e-tests/pages.test.ts
@@ -2,12 +2,24 @@ import { join } from "path";
 import { FrameworkMap } from "frameworks/index";
 import { readJSON } from "helpers/files";
 import { fetch } from "undici";
-import { describe, expect, test, afterEach, beforeEach } from "vitest";
-import { version } from "../package.json";
+import {
+	describe,
+	expect,
+	test,
+	afterEach,
+	beforeEach,
+	afterAll,
+} from "vitest";
 import { deleteProject } from "../scripts/e2eCleanup";
 import { frameworkCliMap } from "../src/frameworks/package.json";
 import { frameworkToTest } from "./frameworkToTest";
-import { isQuarantineMode, keys, runC3, testProjectDir } from "./helpers";
+import {
+	isQuarantineMode,
+	keys,
+	runC3,
+	testDeploymentCommitMessage,
+	testProjectDir,
+} from "./helpers";
 import type { RunnerConfig } from "./helpers";
 import type { TestContext } from "vitest";
 
@@ -22,132 +34,6 @@ type FrameworkTestConfig = RunnerConfig & {
 };
 
 describe.concurrent(`E2E: Web frameworks`, () => {
-	const { getPath, getName, clean } = testProjectDir("pages");
-
-	const getFrameworkNameFromTestContext = (ctx: TestContext) => {
-		const framework = ctx.meta.name.replace(/^Quarantined: /, "");
-		if (!frameworks.includes(framework)) {
-			throw new Error(
-				`Error: Invalid test name, could not detect current framework from it`
-			);
-		}
-		return framework;
-	};
-
-	beforeEach((ctx) => {
-		const framework = getFrameworkNameFromTestContext(ctx);
-		clean(framework);
-	});
-
-	afterEach(async (ctx) => {
-		const framework = getFrameworkNameFromTestContext(ctx);
-		clean(framework);
-
-		// Cleanup the pages project in case we need to retry it
-		const projectName = getName(framework);
-		try {
-			await deleteProject(projectName);
-		} catch (error) {
-			console.error(`Failed to cleanup project: ${projectName}`);
-			console.error(error);
-		}
-	});
-
-	const runCli = async (
-		framework: string,
-		{ argv = [], promptHandlers = [], overrides }: RunnerConfig
-	) => {
-		const projectPath = getPath(framework);
-
-		const args = [
-			projectPath,
-			"--type",
-			"webFramework",
-			"--framework",
-			framework,
-			"--deploy",
-			"--no-open",
-			"--no-git",
-		];
-
-		args.push(...argv);
-
-		const { output } = await runC3({
-			argv: args,
-			promptHandlers,
-			outputPrefix: `[${framework}]`,
-		});
-
-		// Relevant project files should have been created
-		expect(projectPath).toExist();
-		const pkgJsonPath = join(projectPath, "package.json");
-		expect(pkgJsonPath).toExist();
-
-		// Wrangler should be installed
-		const wranglerPath = join(projectPath, "node_modules/wrangler");
-		expect(wranglerPath).toExist();
-
-		// Verify package scripts
-		const frameworkConfig = FrameworkMap[framework];
-
-		const frameworkTargetPackageScripts = {
-			...frameworkConfig.packageScripts,
-		} as Record<string, string>;
-
-		if (overrides && overrides.packageScripts) {
-			// override packageScripts with testing provided scripts
-			Object.entries(overrides.packageScripts).forEach(([target, cmd]) => {
-				frameworkTargetPackageScripts[target] = cmd;
-			});
-		}
-
-		const pkgJson = readJSON(pkgJsonPath);
-		Object.entries(frameworkTargetPackageScripts).forEach(([target, cmd]) => {
-			expect(pkgJson.scripts[target]).toEqual(cmd);
-		});
-
-		return { output };
-	};
-
-	const runCliWithDeploy = async (
-		framework: string,
-		testCommitMessage: boolean
-	) => {
-		const { argv, overrides, promptHandlers, expectResponseToContain } =
-			frameworkTests[framework];
-
-		const { output } = await runCli(framework, {
-			overrides,
-			promptHandlers,
-			argv: [...(argv ?? [])],
-		});
-
-		// Verify deployment
-		const deployedUrlRe =
-			/deployment is ready at: (https:\/\/.+\.(pages|workers)\.dev)/;
-
-		const match = output.match(deployedUrlRe);
-		if (!match || !match[1]) {
-			expect(false, "Couldn't find deployment url in C3 output").toBe(true);
-			return;
-		}
-
-		const projectUrl = match[1];
-
-		const res = await fetch(projectUrl);
-		expect(res.status).toBe(200);
-
-		const body = await res.text();
-		expect(
-			body,
-			`(${framework}) Deployed page (${projectUrl}) didn't contain expected string: "${expectResponseToContain}"`
-		).toContain(expectResponseToContain);
-
-		if (testCommitMessage) {
-			await testDeploymentCommitMessage(getName(framework), framework);
-		}
-	};
-
 	// These are ordered based on speed and reliability for ease of debugging
 	const frameworkTests: Record<string, FrameworkTestConfig> = {
 		astro: {
@@ -255,39 +141,168 @@ describe.concurrent(`E2E: Web frameworks`, () => {
 		},
 	};
 
-	Object.keys(frameworkTests).forEach((framework) => {
-		const config = frameworkTests[framework];
-		if (isQuarantineMode() || config.quarantine) return;
+	const { getPath, getName, clean } = testProjectDir("pages");
 
-		const skip = frameworkToTest && frameworkToTest !== framework;
+	const quarantineStats = { passed: [] as string[], failed: [] as string[] };
 
-		test.skipIf(skip)(
-			framework,
-			async () => {
-				await runCliWithDeploy(
-					framework,
-					frameworkTests[framework].testCommitMessage
-				);
-			},
-			{ retry: 3, timeout: config.timeout || TEST_TIMEOUT }
-		);
+	const getFrameworkNameFromTestContext = (ctx: TestContext) => {
+		const framework = ctx.meta.name.replace(/^Quarantined: /, "");
+		if (!frameworks.includes(framework)) {
+			throw new Error(
+				`Error: Invalid test name, could not detect current framework from it`
+			);
+		}
+		return framework;
+	};
+
+	beforeEach((ctx) => {
+		const framework = getFrameworkNameFromTestContext(ctx);
+		clean(framework);
 	});
 
+	afterEach(async (ctx) => {
+		const framework = getFrameworkNameFromTestContext(ctx);
+		clean(framework);
+
+		// Cleanup the pages project in case we need to retry it
+		const projectName = getName(framework);
+		try {
+			await deleteProject(projectName);
+		} catch (error) {
+			console.error(`Failed to cleanup project: ${projectName}`);
+			console.error(error);
+		}
+	});
+
+	afterAll(() => {
+		if (isQuarantineMode()) {
+			console.log("Quarantine Results");
+			console.log("======================");
+			console.log(`Passed: ${quarantineStats.passed.join(", ")}`);
+			console.log(`Failed: ${quarantineStats.failed.join(", ")}`);
+		}
+	});
+
+	const runCli = async (
+		framework: string,
+		{ argv = [], promptHandlers = [], overrides }: RunnerConfig
+	) => {
+		const projectPath = getPath(framework);
+
+		const args = [
+			projectPath,
+			"--type",
+			"webFramework",
+			"--framework",
+			framework,
+			"--deploy",
+			"--no-open",
+			"--no-git",
+		];
+
+		args.push(...argv);
+
+		const { output } = await runC3({
+			argv: args,
+			promptHandlers,
+			outputPrefix: `[${framework}]`,
+		});
+
+		// Relevant project files should have been created
+		expect(projectPath).toExist();
+		const pkgJsonPath = join(projectPath, "package.json");
+		expect(pkgJsonPath).toExist();
+
+		// Wrangler should be installed
+		const wranglerPath = join(projectPath, "node_modules/wrangler");
+		expect(wranglerPath).toExist();
+
+		// Verify package scripts
+		const frameworkConfig = FrameworkMap[framework];
+
+		const frameworkTargetPackageScripts = {
+			...frameworkConfig.packageScripts,
+		} as Record<string, string>;
+
+		if (overrides && overrides.packageScripts) {
+			// override packageScripts with testing provided scripts
+			Object.entries(overrides.packageScripts).forEach(([target, cmd]) => {
+				frameworkTargetPackageScripts[target] = cmd;
+			});
+		}
+
+		const pkgJson = readJSON(pkgJsonPath);
+		Object.entries(frameworkTargetPackageScripts).forEach(([target, cmd]) => {
+			expect(pkgJson.scripts[target]).toEqual(cmd);
+		});
+
+		return { output };
+	};
+
+	const runCliWithDeploy = async (
+		framework: string,
+		testCommitMessage: boolean
+	) => {
+		const { argv, overrides, promptHandlers, expectResponseToContain } =
+			frameworkTests[framework];
+
+		const { output } = await runCli(framework, {
+			overrides,
+			promptHandlers,
+			argv: [...(argv ?? [])],
+		});
+
+		// Verify deployment
+		const deployedUrlRe =
+			/deployment is ready at: (https:\/\/.+\.(pages|workers)\.dev)/;
+
+		const match = output.match(deployedUrlRe);
+		if (!match || !match[1]) {
+			expect(false, "Couldn't find deployment url in C3 output").toBe(true);
+			return;
+		}
+
+		const projectUrl = match[1];
+
+		const res = await fetch(projectUrl);
+		expect(res.status).toBe(200);
+
+		const body = await res.text();
+		expect(
+			body,
+			`(${framework}) Deployed page (${projectUrl}) didn't contain expected string: "${expectResponseToContain}"`
+		).toContain(expectResponseToContain);
+
+		if (testCommitMessage) {
+			await testDeploymentCommitMessage(getName(framework), framework);
+		}
+	};
+
 	Object.keys(frameworkTests).forEach((framework) => {
-		const config = frameworkTests[framework];
-		if (!isQuarantineMode() || !config.quarantine) return;
+		const { quarantine, timeout, testCommitMessage } =
+			frameworkTests[framework];
 
-		const skip = frameworkToTest && frameworkToTest !== framework;
+		const quarantineSkip = isQuarantineMode() !== (quarantine ?? false);
+		const singleFrameworkSkip =
+			frameworkToTest && frameworkToTest !== framework;
 
-		test.skipIf(skip)(
-			`Quarantined: ${framework}`,
+		test.skipIf(quarantineSkip || singleFrameworkSkip)(
+			framework,
 			async () => {
-				await runCliWithDeploy(
-					framework,
-					frameworkTests[framework].testCommitMessage
-				);
+				try {
+					await runCliWithDeploy(framework, testCommitMessage);
+					if (isQuarantineMode()) {
+						quarantineStats.passed.push(framework);
+					}
+				} catch (error) {
+					if (isQuarantineMode()) {
+						quarantineStats.failed.push(framework);
+					} else {
+						throw error;
+					}
+				}
 			},
-			{ retry: 3, timeout: TEST_TIMEOUT }
+			{ retry: 3, timeout: timeout || TEST_TIMEOUT }
 		);
 	});
 
@@ -295,46 +310,3 @@ describe.concurrent(`E2E: Web frameworks`, () => {
 		await runCli("hono", { argv: ["--wrangler-defaults"] });
 	});
 });
-
-const testDeploymentCommitMessage = async (
-	projectName: string,
-	framework: string
-) => {
-	// Note: we cannot simply run git and check the result since the commit can be part of the
-	//       deployment even without git, so instead we fetch the deployment info from the pages api
-	const response = await fetch(
-		`https://api.cloudflare.com/client/v4/accounts/${process.env.CLOUDFLARE_ACCOUNT_ID}/pages/projects`,
-		{
-			headers: {
-				Authorization: `Bearer ${process.env.CLOUDFLARE_API_TOKEN}`,
-			},
-		}
-	);
-
-	const result = (
-		(await response.json()) as {
-			result: {
-				name: string;
-				latest_deployment?: {
-					deployment_trigger: {
-						metadata?: {
-							commit_message: string;
-						};
-					};
-				};
-			}[];
-		}
-	).result;
-
-	const projectLatestCommitMessage = result.find(
-		(project) => project.name === projectName
-	)?.latest_deployment?.deployment_trigger?.metadata?.commit_message;
-	expect(projectLatestCommitMessage).toMatch(
-		/^Initialize web application via create-cloudflare CLI/
-	);
-	expect(projectLatestCommitMessage).toContain(
-		`C3 = create-cloudflare@${version}`
-	);
-	expect(projectLatestCommitMessage).toContain(`project name = ${projectName}`);
-	expect(projectLatestCommitMessage).toContain(`framework = ${framework}`);
-};
